made another component BulkOrderButton and moved the files there
